refactor: canvasに関わる抽象的なロジックをuseCanvasに切り出す
